refactor: 디렉토리 추가 및 구조 수정
