+++ src/test/java/lotto/model/lotto/BudgetTest.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.lotto;
 
 import lotto.exception.LottoErrorStatus;
 import lotto.exception.LottoException;
